backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=3.14
mssql-django>=1.4
python-jose[cryptography]>=3.3
pyodbc>=4.0